Fix kernel crash on camera subdevice removale
When the camera subdevice is deregistered (e.g. caused by unloading of the camera driver module) then the kernel crashes with an null pointer error. The error occurs, because in "tegra_channel_remove_subdevices" the video device pointer is set to NULL and is afterwards accessed in "tegra_channel_cleanup_video" to free the "prio" field. The issue is fixed by implementing a proper release function for the video device.